# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/studios/current": {
      "get": {
        "operationId": "StudiosController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/onboarding": {
      "get": {
        "operationId": "StudiosController_onboarding",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/onboarding/profile": {
      "patch": {
        "operationId": "StudiosController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudioProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/onboarding/operation": {
      "patch": {
        "operationId": "StudiosController_updateOperation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudioOperationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/onboarding/plans": {
      "post": {
        "operationId": "StudiosController_saveInitialPlans",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInitialPlansDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/branding": {
      "patch": {
        "operationId": "StudiosController_updateBranding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudioBrandingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/logo/uploads": {
      "post": {
        "operationId": "StudiosController_requestLogoUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestStudioLogoUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/logo/{id}/confirm": {
      "post": {
        "operationId": "StudiosController_confirmLogoUpload",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
    "/studios/logo": {
      "delete": {
        "operationId": "StudiosController_removeLogo",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "studios"
        ]
      }
    },
...
    "/students": {
      "get": {
        "operationId": "StudentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      },
      "post": {
        "operationId": "StudentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/students/duplicates/check": {
      "get": {
        "operationId": "StudentsController_duplicates",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/students/quick": {
      "post": {
        "operationId": "StudentsController_createQuick",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuickStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/students/{id}": {
      "get": {
        "operationId": "StudentsController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      },
      "patch": {
        "operationId": "StudentsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/students/{id}/archive": {
      "post": {
        "operationId": "StudentsController_archive",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
...
  },
  "info": {
    "title": "pilates-manager API",
    "description": "Backend API for Pilates studio management",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "UpdateStudioProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateStudioOperationDto": {
        "type": "object",
        "properties": {}
      },
      "SaveInitialPlansDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateStudioBrandingDto": {
        "type": "object",
        "properties": {}
      },
      "RequestStudioLogoUploadDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateQuickStudentDto": {
        "type": "object",
        "properties": {}
      },
      "CreateStudentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}